refactor(docker/images): move images list proxy to proxy folder
