+++ app/react/components/ImageConfigFieldset/SimpleForm.tsx
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { useMemo } from 'react';
 
 import DockerIcon from '@/assets/ico/vendor/docker.svg?c';
-import { useImages } from '@/react/docker/images/queries/useImages';
+import { useImages } from '@/react/docker/proxy/queries/images/useImages';
 import {
   imageContainsURL,
   getUniqueTagListFromImages,
