+++ app/react/docker/proxy/queries/images/queryKeys.ts
@@ -1,9 +1,9 @@
 import { EnvironmentId } from '@/react/portainer/environments/types';
 
-import { queryKeys as dockerQueryKeys } from '../../queries/utils';
+import { queryKeys as proxyQueryKeys } from '../query-keys';
 
 export const queryKeys = {
   base: (environmentId: EnvironmentId) =>
-    [dockerQueryKeys.root(environmentId), 'images'] as const,
+    [proxyQueryKeys.base(environmentId), 'images'] as const,
   list: (environmentId: EnvironmentId) => queryKeys.base(environmentId),
 };
